<Learning>(<fundamentals>): Finished Data Type.
The data type portion of the fundamentals review is finished.
Various data types were reviewed, such as scalar types like integers, floats, booleans, chars,
as well as compound types like tuples and arrays. In regards to the ladder, both how such memory is stored
as well as accessing specific elements of the types was discussed.