<Learning>(<fundamentals>): Finished functions.
I finished the functions portion of the fundamentals review. I covered how functions are instantiated and called,
as well as how to make use of parameters (arguments) for function input and output (calling). Finally, I learned the
difference between a statement, which is a piece of code that perform actions but do not return a value, and expressions,
which evaluate a resultant value.