<Learning>(<fundamentals>)Finished all review sections. <Practice>(<Fundamentals>)Completed 3 practice challenges.
The fundamentals portion of the Rust review has been completed. I finished out the "learning" with control control_flow
(i.e. loops, if/else statements) and functions. I also completed three challenge prompts that were given at the end of the
fundamentals chapter. The first of which was to create a Fahrenheit to Celsius and vice versa converter. The second challenge
was a program to determine the nth number in a Fibonacci sequence. The third and final program was to iteratively build
the "Twelve Days of Christmas" song.