<Learning>(<Ownership>): Completed 4.1 Ownership
Learned about one of the things that makes Rust unique, Ownership. Memory in Rust is still stored on the
stack and the heap, but rather than a Garbage Collector to clean up unmaintained data left of the heap,
it uses Ownership to automatically drop variables that are no longer considered in scope. In other words,
when allocating data to the heap, a memory allocator is used, and when the data is out of scope (i.e. function ends),
the variable is dropped. There is some unique behavior associated with this, namely copying, or better referenced as
moving data around. See notes in script for more information.